+++ phpmyfaq/admin/assets/src/configuration/index.ts
@@ -2,5 +2,6 @@ export * from './configuration';
 export * from './elasticsearch';
 export * from './forms';
 export * from './instance';
+export * from './opensearch';
 export * from './stopwords';
 export * from './upgrade';